requirements.txt
psycopg[binary]
pandas
datasets
python-dotenv